task2 upd
consoe is working and have been tested

+++ task2/src/main/java/filemanager/FileManagerImpl.java
@@ -1,6 +1,6 @@
 package filemanager;
 
-
+import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Pattern;
@@ -8,7 +8,7 @@ import java.util.regex.Pattern;
 public class FileManagerImpl implements FileManager {
 
 
-    public List<String> searchByPattern(String path, String regex) {
+    public List<String> search(String path, String regex) throws IllegalArgumentException {
         List<String> resultList = new ArrayList<String>();
         PathTreeNode treeNode = FileUtils.fillFileTree(new PathTreeNode(path));
         Pattern pattern = Pattern.compile(regex);
@@ -18,7 +18,8 @@ public class FileManagerImpl implements FileManager {
             while (currentLvl.size() != 0) {
                 for (TreeNode<String> child : currentLvl) {
                     if (child.getPath() != null) {
-                        if (pattern.matcher(child.getPath()).matches()) {
+                        String s = child.getPath();
+                        if (pattern.matcher(s).matches()) {
                             resultList.add(child.getFullPath());
                         }
                     }
@@ -33,7 +34,11 @@ public class FileManagerImpl implements FileManager {
         return resultList;
     }
 
-    public void copyFiles(String sourcePath, String destinationPath) {
+    public void copy(String sourcePath, String destinationPath) throws IllegalArgumentException {
+        File dest = new File(destinationPath);
+        if (!dest.exists()) {
+            throw new IllegalArgumentException();
+        }
         PathTreeNode treeNode = FileUtils.fillFileTree(new PathTreeNode(sourcePath));
         PathTreeNode copyTreeNode = (PathTreeNode) treeNode.copy();
         copyTreeNode.setData(destinationPath);
@@ -56,32 +61,38 @@ public class FileManagerImpl implements FileManager {
     }
 
 
-    public void moveFiles(String sourcePath, String destinationPath) {
+    public void move(String sourcePath, String destinationPath) throws IllegalArgumentException {
+        File dest = new File(destinationPath);
+        if (!dest.exists()) {
+            throw new IllegalArgumentException();
+        }
         PathTreeNode treeNode = FileUtils.fillFileTree(new PathTreeNode(sourcePath));
         if (sourcePath.charAt(0) == destinationPath.charAt(0)) {
             FileUtils.changePath(sourcePath, destinationPath);
-        }
-        PathTreeNode copyTreeNode = (PathTreeNode) treeNode.copy();
-        copyTreeNode.setData(destinationPath);
+        } else {
+            PathTreeNode copyTreeNode = (PathTreeNode) treeNode.copy();
+            copyTreeNode.setData(destinationPath);
 
-        if (!treeNode.isLeaf()) {
-            List<TreeNode<String>> currentLvl = treeNode.getChildren();
-            List<TreeNode<String>> nextLvl = new ArrayList<TreeNode<String>>();
-            while (currentLvl.size() != 0) {
-                for (TreeNode<String> child : currentLvl) {
-                    if (child.isLeaf()) {
-                        FileUtils.copyFile(child.getFullPath(), destinationPath + "\\" + child.getSubPath());
-                        FileUtils.deleteFile(child.getFullPath());
-                    } else {
-                        nextLvl.addAll(child.getChildren());
-                        FileUtils.createDirectory(destinationPath, child.getPath());
+            if (!treeNode.isLeaf()) {
+                List<TreeNode<String>> currentLvl = treeNode.getChildren();
+                List<TreeNode<String>> nextLvl = new ArrayList<TreeNode<String>>();
+                while (currentLvl.size() != 0) {
+                    for (TreeNode<String> child : currentLvl) {
+                        if (child.isLeaf()) {
+                            FileUtils.copyFile(child.getFullPath(), destinationPath + "\\" + child.getSubPath());
+                            FileUtils.deleteFile(child.getFullPath());
+                        } else {
+                            nextLvl.addAll(child.getChildren());
+                            FileUtils.createDirectory(destinationPath, child.getPath());
+                        }
                     }
+                    currentLvl = nextLvl;
+                    nextLvl = new ArrayList<TreeNode<String>>();
                 }
-                currentLvl = nextLvl;
-                nextLvl = new ArrayList<TreeNode<String>>();
             }
         }
     }
+
 /*
     private void traverseTree(Action action, TreeNode<String> treeNode, Pattern pattern, List<String> resultList ) {
